feat: update tests on frame support pallet_outer

+++ substrate/frame/support/test/tests/pallet_outer_enums_implicit.rs
@@ -21,7 +21,7 @@ mod common;
 
 use common::outer_enums::{pallet, pallet2};
 
-pub type Header = sp_runtime::generic::Header<u32, sp_runtime::traits::BlakeTwo256>;
+pub type Header = sp_runtime::generic::Header<u64, sp_runtime::traits::BlakeTwo256>;
 pub type Block = sp_runtime::generic::Block<Header, UncheckedExtrinsic>;
 pub type UncheckedExtrinsic = sp_runtime::generic::UncheckedExtrinsic<u32, RuntimeCall, (), ()>;
